agregando una condicional a procedimiento almacado para no eliminar al categoria por defecto

+++ sql/StoreProcedures.sql
@@ -1,11 +1,13 @@
 DELIMITER //
-  CREATE PROCEDURE sp_eliminar_categoria(IN p_id_categoria INT)
-  BEGIN
-      UPDATE productos
-      SET categoria_id = 1
-      WHERE categoria_id = p_id_categoria;
-      
-      DELETE FROM categorias
-      WHERE id = p_id_categoria;
-  END //
+	CREATE PROCEDURE sp_eliminar_categoria(IN p_id_categoria INT)
+	BEGIN
+			IF p_id_categoria <> 1 THEN
+					UPDATE productos 
+					SET categoria_id = 5 
+					WHERE categoria_id = p_id_categoria;
+
+					DELETE FROM categorias 
+					WHERE id = p_id_categoria;					
+			END IF;
+	END //
 DELIMITER ;
